% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_functions.R
\name{model_results}
\alias{model_results}
\title{Model results}
\usage{
model_results(data, n_boot)
}
\arguments{
\item{data}{the data set used to fit the model}

\item{n_boot}{the number of bootstraping samples to use}
}
\value{
A dataframe with important information about the predictors like point estimates, tstatistic, standard error and confidence intervals.
}
\description{
Model results
}
\examples{
##### Epilepsy #####

## Load data
epilepsy <- read_csv("epilepsy.csv")

## Sum separate observations for each patient in the after period
epilepsy <- epilepsy \%>\%
  group_by(id,treat,expind,age) \%>\%
  summarize(seizures = sum(seizures),
            .groups = "drop")

## Fit model to epilepsy data. Fixed effects in the model are:
##  (Intercept) -- the intercept
##  age -- age as a continuous predictor
##  expind -- a categorical variable with two levels (0 for before and 1 for after)
##  expind:treat -- a categorical variable with two levels (1 for observations from individuals on the drug in the after period and 0 otherwise)

results <- model_results(epilepsy, "epilepsy")
}
