% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_functions.R
\name{pValue_age}
\alias{pValue_age}
\title{P Value of the predictor age}
\usage{
pValue_age(data, t_value, n_boot)
}
\arguments{
\item{data}{the data set used to fit the model}

\item{t_value}{statistic of predictor when the model was fitted}

\item{n_boot}{the number of bootstraping samples to use}
}
\value{
The pvalue of the hypothesis predictor age equal to 0
}
\description{
P Value of the predictor age
}
\examples{
##### Epilepsy #####

## Load data
epilepsy <- read_csv("epilepsy.csv")

## Sum separate observations for each patient in the after period
epilepsy <- epilepsy \%>\%
  group_by(id,treat,expind,age) \%>\%
  summarize(seizures = sum(seizures),
            .groups = "drop")

pValue_age <- pValue_age(epilepsy,1.16,1000)
}
