% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_functions.R
\name{run_model}
\alias{run_model}
\title{Final Project Model Fitting}
\usage{
run_model(data, example = "tortoise")
}
\arguments{
\item{data}{the data set for your example}

\item{example}{the name of your example: one of "culcita", "ctsib", "epilepsy", or "tortoise"}
}
\value{
A list with summary statistics from the fitted model.
}
\description{
Final Project Model Fitting
}
