% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_functions.R
\name{model_results}
\alias{model_results}
\title{Model results}
\usage{
model_results(data, n_boot)
}
\arguments{
\item{data}{the data set used to fit the model}

\item{n_boot}{the number of bootstraping samples to use}
}
\value{
A dataframe with important information about the predictors like point estimates, tstatistic, standard error and confidence intervals.
}
\description{
Model results
}
